requirements.txt
backcall==0.1.0
decorator==4.3.0
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.8.2
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
tabulate==0.8.2
traitlets==4.3.2
wcwidth==0.1.7
